codeswiftr/requirements.txt
asgiref==3.2.5
autopep8==1.4.4
beautifulsoup4==4.8.2
certifi==2019.11.28
coverage==5.0.4
Django==3.0.3
django-bootstrap4==1.1.1
django-braces==1.14.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Faker==4.0.2
pycodestyle==2.5.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
text-unidecode==1.3
